src/app/requirements.txt
Flask
mysql-connector
tweepy
flask_mail
kafka-python